Added to build image with own WAR o downloaded

+++ entrypoint.sh
@@ -2,23 +2,27 @@
 set -e
 
 # wait upto 30 seconds for the database to start before connecting
-/wait-for-it.sh $DB_HOST:$DB_PORT -t 30
+#/wait-for-it.sh $DB_HOST:$DB_PORT -t 30
 
 # check if we need to bootstrap the JasperServer
 if [ ! -d "$CATALINA_HOME/webapps/jasperserver" ]; then
     pushd /usr/src/jasperreports-server/buildomatic
-    
+
     # only works for Postgres or MySQL
-    cp sample_conf/${DB_TYPE}_master.properties default_master.properties
-    sed -i -e "s|^appServerDir.*$|appServerDir = $CATALINA_HOME|g; s|^dbHost.*$|dbHost=$DB_HOST|g; s|^dbPort.*$|dbPort=$DB_PORT|g; s|^dbUsername.*$|dbUsername=$DB_USER|g; s|^dbPassword.*$|dbPassword=$DB_PASSWORD|g" default_master.properties
+    if [ -e "/jasperserver-conf/default_master.properties" ]; then
+      cp /jasperserver-conf/default_master.properties default_master.properties
+    else
+      cp sample_conf/${DB_TYPE}_master.properties default_master.properties
+      sed -i -e "s|^appServerDir.*$|appServerDir = $CATALINA_HOME|g; s|^dbHost.*$|dbHost=$DB_HOST|g; s|^dbPort.*$|dbPort=$DB_PORT|g; s|^dbUsername.*$|dbUsername=$DB_USER|g; s|^dbPassword.*$|dbPassword=$DB_PASSWORD|g" default_master.properties
+    fi
 
     # run the minimum bootstrap script to initial the JasperServer
     ./js-ant create-js-db || true #create database and skip it if database already exists
-    ./js-ant init-js-db-ce 
-    ./js-ant import-minimal-ce 
+    ./js-ant init-js-db-ce
+    ./js-ant import-minimal-ce
     ./js-ant deploy-webapp-ce
-    
-    
+
+
     # import any export zip files from another JasperServer
 
     shopt -s nullglob # handle case if no zip files found
@@ -27,7 +31,7 @@ if [ ! -d "$CATALINA_HOME/webapps/jasperserver" ]; then
     for f in $IMPORT_FILES
     do
       echo "Importing $f..."
-      ./js-import.sh --input-zip $f 
+      ./js-import.sh --input-zip $f
     done
